fix: this page is no longer used and can be removed
